requirements.txt
attrs==19.3.0
Automat==20.2.0
cffi==1.14.0
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
lxml==4.5.1
parsel==1.6.0
pkg-resources==0.0.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
queuelib==1.5.0
Scrapy==2.1.0
service-identity==18.1.0
six==1.15.0
Twisted==20.3.0
w3lib==1.22.0
zope.interface==5.1.0
attrs==19.3.0
Automat==20.2.0
autopep8==1.5.2
boto3==1.13.15
botocore==1.16.15
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
dateparser==0.7.4
docutils==0.15.2
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
jmespath==0.10.0
lxml==4.5.1
numpy==1.18.4
pandas==1.0.3
parsel==1.6.0
pdfminer.six==20200104
pdfplumber==0.5.20
Pillow==7.1.2
pkg-resources==0.0.0
price-parser==0.3.3
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.7
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
queuelib==1.5.0
regex==2020.5.14
s3transfer==0.3.3
Scrapy==2.1.0
service-identity==18.1.0
six==1.15.0
sortedcontainers==2.1.0
Twisted==20.3.0
tzlocal==2.1
unicode-slugify==0.1.3
unicodecsv==0.14.1
Unidecode==1.1.1
urllib3==1.25.9
w3lib==1.22.0
Wand==0.6.1
zope.interface==5.1.0
